+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 7.0
 - Update MoTranslator from v3.4 to v4.0 in Warehouse.php, Help.fnc.php & classes/MoTranslator/*
 - Fix 'School' translation when using MoTranslator in Schedule.inc.php & rosariosis.po
 - Fix '%s Handbook' translation when using MoTranslator in Help.php
+- CSS fix align "+" New Event icon to bottom in Calendar.php, CalendarDay.inc.php, stylesheet.css & zreponsive.css
 
 Changes in 6.9.3
 ----------------